move loop node auxfuns into node_aux.py
