+++ endf_parserpy/compiler/cpp_types/cpp_type_matrix2d/assign.py
@@ -24,13 +24,15 @@ from endf_parserpy.compiler.variable_management import (
     get_var_types,
 )
 from endf_parserpy.compiler.node_checks import is_loop
-from ..cpp_varaux import (
-    check_variable,
-    has_loopvartype,
+from endf_parserpy.compiler.node_aux import (
     get_loopvar,
     get_loop_body,
     get_loop_start,
     get_loop_stop,
+)
+from ..cpp_varaux import (
+    check_variable,
+    has_loopvartype,
     init_local_var_from_global_var,
 )
 from ..cpp_dtype_aux import map_dtype
